# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ResponseHTTP:
    properties:
      data: {}
      details:
        type: string
      success:
        type: boolean
    type: object
  common.ResponsePagination:
    properties:
      data: {}
      details:
        type: string
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  controllers.UserPassword:
    properties:
      email:
        example: someone@domain.com
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.Endpoint:
    description: App type information
    properties:
      description:
        type: string
      feature_id:
        type: number
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      route_path:
        type: string
    type: object
...
  models.Feature:
    description: App type information
    properties:
      active:
        type: boolean
      description:
        type: string
      endpoints:
        items:
          $ref: '#/definitions/models.Endpoint'
        type: array
      id:
        type: integer
      name:
        type: string
      role:
        type: number
    type: object
...
  models.Page:
    description: App type information
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  models.Role:
    description: App type information
    properties:
      active:
        type: boolean
      app:
        type: number
      description:
        type: string
      features:
        items:
          $ref: '#/definitions/models.Feature'
        type: array
      id:
        type: integer
      name:
        type: string
      pages:
        items:
          $ref: '#/definitions/models.Page'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.User:
    description: App type information
    properties:
      date_registered:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
  models.UserGet:
    description: UserGet type information
    properties:
      date_registered:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
  models.UserPost:
    description: UserPost type information
    properties:
      disabled:
        type: boolean
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: This is blue-admin API OPENAPI Documentation.
  termsOfService: http://swagger.io/terms/
  title: Swagger blue-admin API
  version: "0.1"
paths:
...
  /roleuser/{role_id}/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete Role User
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.UserPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add User
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Add Role User
      parameters:
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add User to Role
      tags:
      - RoleUsers
  /user:
    get:
      consumes:
      - application/json
      description: Get Users
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Add User
      parameters:
      - description: Add User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.UserPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new User
      tags:
      - Users
  /user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove user by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove User by ID
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.UserGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get User by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Patch User
      parameters:
      - description: Patch User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserPost'
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.UserPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch User
      tags:
      - Users
...
  /users:
    put:
      consumes:
      - application/json
      description: Put User
      parameters:
      - description: Password User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserPassword'
      - description: Reset Password
        in: query
        name: reset
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.UserGet'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Put User
      tags:
      - Users
  /users/{user_id}:
    put:
      consumes:
      - application/json
      description: Activate/Deactivate User
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Disabled
        in: query
        name: status
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Activate/Deactivate User
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: X-APP-TOKEN
    type: apiKey
  Refresh:
    description: Description for what is this security definition being used
    in: header
    name: X-REFRESH-TOKEN
    type: apiKey
swagger: "2.0"